rest_server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
grpcio==1.60.0
protobuf==4.21.12
